api_test/requirements.txt
requests
pytest
pytest-html
allure-pytest
PyMySQL
Pillow
PyYAML